Solution to Calculate Average Temperature of a Week

+++ Calculate Average Temperature of a Week@@67877d61dc107486ad5c274d/code.c
@@ -2,13 +2,13 @@
 
 typedef struct temperature{
     char day[50];
-    int temp;
+    double temp;
 } temperature;
 
 int main(){
     temperature t[7];
     for(int i=0;i<=7;i++){
-        scanf("%s %d",t[i].day,&t[i].temp);
+        scanf("%s %lf",t[i].day,&t[i].temp);
     }
     printf("Average Temperature: %.2f",s[3].temp);
     return 0;
